Implement updateFutureVars (updateFutureArcs helper)
My poor bren... These are very few lines of code, but boy did they
require thinking through. I'm still not sure they're correct, but I (a)
don't want to try to work out a type-encoding for proving/guaranteeing
that they are, and (b) am perfectly happy with debugging things if this
fails (although future-me will probably hate me for that).

This is an intermediary commit before I fully wire things together for
`reviseFutureArcs`.